#242 Adding confirmation deletion message for mood journal

+++ client/app/moodjournal/viewMoodJournalsPage.tsx
@@ -12,6 +12,7 @@ import { useEffect, useState } from 'react';
 import { formatDate, formatMilitaryTime } from '../helpers/utils/datetimeformat';
 import { useAuth } from '../contexts/AuthContext';
 import Header from '../components/Header';
+import Swal from 'sweetalert2';
 
 export default function ViewMoodJournalsPage() {
     const logger = require('../../logger');
@@ -57,10 +58,28 @@ export default function ViewMoodJournalsPage() {
     }
 
     async function deleteMoodJournals(moodJournalId: string){
-      const deleteresult = await deleteMoodJournal(moodJournalId);   
-      const newData = moodJournal && moodJournal.filter((item: { id: string; }) => item.id!=moodJournalId);
-      setMoodJournal(newData);
-      router.push('/moodjournal');
+      Swal.fire({
+        title: "Are you sure you want to delete this mood journal entry?",
+        text: "You will not be able to retrieve it later",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Delete",
+        position: "bottom",
+      }).then(async (result: { isConfirmed: any; }) => {
+        if (result.isConfirmed) {
+          const deleteresult = await deleteMoodJournal(moodJournalId);   
+          const newData = moodJournal && moodJournal.filter((item: { id: string; }) => item.id!=moodJournalId);
+          setMoodJournal(newData);
+          router.push('/moodjournal');
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your mood journal entry has been deleted.",
+            icon: "success"
+          });    
+        }
+     }); 
+    
     }
 
     const handleClick = (moodJournalID: string) => {
